+++ firmware/telemetry_display/ssd1306.c
@@ -1,191 +1,28 @@
-/*
- * Copyright (c) 2021 Raspberry Pi (Trading) Ltd.
- *
- * SPDX-License-Identifier: BSD-3-Clause
-*/
-
 #include <string.h>
-#include <stdlib.h>
-#include <ctype.h>
 #include "hardware/gpio.h"
 #include "hardware/i2c.h"
-#include "ssd1306_font.h"
 #include "ssd1306.h"
+#include "ssd1306_i2c.h"
 
 /*
-   The SSD1306 is an OLED/PLED driver chip, capable of driving displays up to
-   128x64 pixels.
-
-   NOTE: Ensure the device is capable of being driven at 3.3v NOT 5v. The Pico
-   GPIO (and therefore I2C) cannot be used at 5v.
-
-   You will need to use a level shifter on the I2C lines if you want to run the
-   board at 5v.
-
-   Connections on Raspberry Pi Pico board, other boards may vary.
-
-   GPIO PICO_DEFAULT_I2C_SDA_PIN (on Pico this is GP4 (pin 6)) -> SDA on display
-   board
-   GPIO PICO_DEFAULT_I2C_SCL_PIN (on Pico this is GP5 (pin 7)) -> SCL on
-   display board
-   3.3v (pin 36) -> VCC on display board
-   GND (pin 38)  -> GND on display board
+https://en.cppreference.com/cpp/string/byte/memset
 */
-
-
-void calc_render_area_buflen(struct render_area *area) {
-    // calculate how long the flattened buffer will be for a render area
-    area->buflen = (area->end_col - area->start_col + 1) * (area->end_page - area->start_page + 1);
-}
-
-
-void SSD1306_send_cmd(uint8_t cmd) {
-    // I2C write process expects a control byte followed by data
-    // this "data" can be a command or data to follow up a command
-    // Co = 1, D/C = 0 => the driver expects a command
-    uint8_t buf[2] = {0x80, cmd};
-    i2c_write_blocking(i2c_default, SSD1306_I2C_ADDR, buf, 2, false);
-}
-
-void SSD1306_send_cmd_list(uint8_t *buf, int num) {
-    for (int i=0;i<num;i++)
-        SSD1306_send_cmd(buf[i]);
-}
-
-void SSD1306_send_buf(uint8_t buf[], int buflen) {
-    // in horizontal addressing mode, the column address pointer auto-increments
-    // and then wraps around to the next page, so we can send the entire frame
-    // buffer in one go!
-    // copy our frame buffer into a new buffer because we need to add the control byte
-    // to the beginning
-
-    uint8_t *temp_buf = malloc(buflen + 1);
-
-    temp_buf[0] = 0x40;
-    memcpy(temp_buf+1, buf, buflen);
-
-    i2c_write_blocking(i2c_default, SSD1306_I2C_ADDR, temp_buf, buflen + 1, false);
-
-    free(temp_buf);
-}
-
-void SSD1306_init() {
-    // Some of these commands are not strictly necessary as the reset
-    // process defaults to some of these but they are shown here
-    // to demonstrate what the initialization sequence looks like
-    // Some configuration values are recommended by the board manufacturer
-
-    uint8_t cmds[] = {
-        SSD1306_SET_DISP,               // set display off
-        /* memory mapping */
-        SSD1306_SET_MEM_MODE,           // set memory address mode 0 = horizontal, 1 = vertical, 2 = page
-        0x00,                           // horizontal addressing mode
-        /* resolution and layout */
-        SSD1306_SET_DISP_START_LINE,    // set display start line to 0
-        SSD1306_SET_SEG_REMAP | 0x01,   // set segment re-map, column address 127 is mapped to SEG0
-        SSD1306_SET_MUX_RATIO,          // set multiplex ratio
-        SSD1306_HEIGHT - 1,             // Display height - 1
-        SSD1306_SET_COM_OUT_DIR | 0x08, // set COM (common) output scan direction. Scan from bottom up, COM[N-1] to COM0
-        SSD1306_SET_DISP_OFFSET,        // set display offset
-        0x00,                           // no offset
-        SSD1306_SET_COM_PIN_CFG,        // set COM (common) pins hardware configuration. Board specific magic number.
-                                        // 0x02 Works for 128x32, 0x12 Possibly works for 128x64. Other options 0x22, 0x32
-#if ((SSD1306_WIDTH == 128) && (SSD1306_HEIGHT == 32))
-        0x02,
-#elif ((SSD1306_WIDTH == 128) && (SSD1306_HEIGHT == 64))
-        0x12,
-#else
-        0x02,
-#endif
-        /* timing and driving scheme */
-        SSD1306_SET_DISP_CLK_DIV,       // set display clock divide ratio
-        0x80,                           // div ratio of 1, standard freq
-        SSD1306_SET_PRECHARGE,          // set pre-charge period
-        0xF1,                           // Vcc internally generated on our board
-        SSD1306_SET_VCOM_DESEL,         // set VCOMH deselect level
-        0x30,                           // 0.83xVcc
-        /* display */
-        SSD1306_SET_CONTRAST,           // set contrast control
-        0xFF,
-        SSD1306_SET_ENTIRE_ON,          // set entire display on to follow RAM content
-        SSD1306_SET_NORM_DISP,           // set normal (not inverted) display
-        SSD1306_SET_CHARGE_PUMP,        // set charge pump
-        0x14,                           // Vcc internally generated on our board
-        SSD1306_SET_SCROLL | 0x00,      // deactivate horizontal scrolling if set. This is necessary as memory writes will corrupt if scrolling was enabled
-        SSD1306_SET_DISP | 0x01, // turn display on
-    };
-
-    SSD1306_send_cmd_list(cmds, count_of(cmds));
-}
-
-void render(uint8_t *buf, struct render_area *area) {
-    // update a portion of the display with a render area
-    uint8_t cmds[] = {
-        SSD1306_SET_COL_ADDR,
-        area->start_col,
-        area->end_col,
-        SSD1306_SET_PAGE_ADDR,
-        area->start_page,
-        area->end_page
-    };
-
-    SSD1306_send_cmd_list(cmds, count_of(cmds));
-    SSD1306_send_buf(buf, area->buflen);
-}
-
-int GetFontIndex(uint8_t ch) {
-    if (ch >= 'A' && ch <='Z') {
-        return  ch - 'A' + 1;
-    }
-    else if (ch >= '0' && ch <='9') {
-        return  ch - '0' + 27;
-    }
-    else return  0; // Not got that char so space.
-}
-
-void WriteChar(uint8_t *buf, int16_t x, int16_t y, uint8_t ch) {
-    if (x > SSD1306_WIDTH - 8 || y > SSD1306_HEIGHT - 8)
-        return;
-
-    // For the moment, only write on Y row boundaries (every 8 vertical pixels)
-    y = y/8;
-
-    ch = toupper(ch);
-    int idx = GetFontIndex(ch);
-    int fb_idx = y * 128 + x;
-
-    for (int i=0;i<8;i++) {
-        buf[fb_idx++] = font[idx * 8 + i];
-    }
-}
-
-void WriteString(uint8_t *buf, int16_t x, int16_t y, char *str) {
-    // Cull out any string off the screen
-    if (x > SSD1306_WIDTH - 8 || y > SSD1306_HEIGHT - 8)
-        return;
-
-    while (*str) {
-        WriteChar(buf, x, y, *str++);
-        x+=8;
-    }
-}
-
-/*
-https://learn.microsoft.com/pt-br/cpp/c-runtime-library/reference/memset-wmemset?view=msvc-170
-*/
-void ssd1306_clear(uint8_t *buffer) {
+void ssd1306Clear(uint8_t *buffer) {
     memset(buffer, 0, SSD1306_BUF_LEN);
 }
 
-void ssd1306_setup(void) {
-    i2c_init(SSD1306_I2C_PORT, SSD1306_I2C_CLK * 1000);
-    gpio_set_function(SSD1306_I2C_SDA, GPIO_FUNC_I2C);
-    gpio_set_function(SSD1306_I2C_SCL, GPIO_FUNC_I2C);
-    gpio_pull_up(SSD1306_I2C_SDA);
-    gpio_pull_up(SSD1306_I2C_SCL);
+// These are the board specific pins
+void ssd1306Setup(void) {
+    i2c_init(i2c_default, SSD1306_I2C_CLK * 1000);
+    gpio_set_function(PICO_DEFAULT_I2C_SDA_PIN, GPIO_FUNC_I2C);
+    gpio_set_function(PICO_DEFAULT_I2C_SCL_PIN, GPIO_FUNC_I2C);
+    gpio_pull_up(PICO_DEFAULT_I2C_SDA_PIN);
+    gpio_pull_up(PICO_DEFAULT_I2C_SCL_PIN);
+
 }
 
-void ssd1306_render_full(uint8_t *buffer) {
+// Requires the target board to define a default I2C (see pico-sdk board header / i2c_default)
+void ssd1306RenderFull(uint8_t *buffer) {
     struct render_area frame_area = {
         .start_col = 0,
         .end_col = SSD1306_WIDTH - 1,